requirements.txt
boto3==1.4.7
botocore==1.7.48
certifi==2019.6.16
chardet==3.0.4
cryptography==3.4.7
decorator==4.4.2
docutils==0.15.2
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.1.0
Flask-API==2.0
Flask-Caching==1.7.2
Flask-Cors==3.0.7
Flask-MySQL==1.4.0
Flask-ReCaptcha==0.4.2
Flask-SQLAlchemy==2.2
future==0.18.2
futures==3.1.1
gunicorn==19.7.1
http-ece==1.1.0
httplib2==0.13.1
idna==2.5
itsdangerous==0.24
jdcal==1.4.1
Jinja2==3.1.2
jmespath==0.9.4
MarkupSafe==1.1.0
num2words==0.5.7
oauth2client==4.1.3